// src/app/shared/dto/in-list-show.dto.ts
export interface InListShowDto {
  backdrop_path: string
  id: number
  original_title?: string
  overview: string
  poster_path: string
  media_type?: string
  adult: boolean
  title?: string
  name?: string
  original_name: string
  original_language: string
  genre_ids: number[]
  popularity: number
  release_date?: string
  first_air_date?: string
  video: boolean
  vote_average: number
  vote_count: number
  runtime?: number
  number_of_episodes?: number
  number_of_seasons?: number
  genres?: any[]
  season?: number
  episode?: number
  trailer_key?: string
}